+++ src/main/java/com/telepathicgrunt/repurposedstructures/biomeinjection/Mansions.java
@@ -18,14 +18,14 @@ public final class Mansions {
             event.addStructure(RSConfiguredStructures.MANSION_BIRCH);
         }
 
-        else if (RSMansionsConfig.mansionJungleAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionJungleAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_JUNGLE.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.JUNGLE)))
         {
             event.addStructure(RSConfiguredStructures.MANSION_JUNGLE);
         }
 
-        else if (RSMansionsConfig.mansionOakAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionOakAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_OAK.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.FOREST) &&
                         !BiomeSelection.hasNameTemp(event, "birch", "dark", "spooky", "dead", "haunted")))
@@ -33,7 +33,7 @@ public final class Mansions {
             event.addStructure(RSConfiguredStructures.MANSION_OAK);
         }
 
-        else if (RSMansionsConfig.mansionSavannaAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionSavannaAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_SAVANNA.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.SAVANNA) &&
                         !BiomeSelection.isBiomeTemp(event, Biomes.SAVANNA_PLATEAU)))
@@ -41,22 +41,22 @@ public final class Mansions {
             event.addStructure(RSConfiguredStructures.MANSION_SAVANNA);
         }
 
-        else if (RSMansionsConfig.mansionTaigaAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionTaigaAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_TAIGA.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.TAIGA) && event.biome.getPrecipitation() != Biome.Precipitation.SNOW &&
-                        !BiomeSelection.hasNameTemp(event, "giant", "redwood", "snow", "ice", "icy", "glacier", "frozen")))
+                        !BiomeSelection.hasNameTemp(event, "giant", "redwood", "snow", "ice", "icy", "glacier", "glacial", "frozen")))
         {
             event.addStructure(RSConfiguredStructures.MANSION_TAIGA);
         }
 
-        else if (RSMansionsConfig.mansionDesertAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionDesertAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_DESERT.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.DESERT)))
         {
             event.addStructure(RSConfiguredStructures.MANSION_DESERT);
         }
 
-        else if (RSMansionsConfig.mansionSnowyAverageChunkDistance.get() != 1001 &&
+        if (RSMansionsConfig.mansionSnowyAverageChunkDistance.get() != 1001 &&
                 BiomeSelection.isBiomeAllowedTemp(event, RSStructures.MANSION_SNOWY.get(),
                         () -> BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.ICY) ||
                         (BiomeSelection.haveCategoriesTemp(event, Biome.BiomeCategory.TAIGA) && event.biome.getPrecipitation() == Biome.Precipitation.SNOW)))