+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSBiomeDimConfig.java
@@ -93,11 +93,13 @@ public class RSBiomeDimConfig implements Config {
             addEntries(allowedDimensions, "repurposed_structures:well_forest", "lotr:middle_earth");
             addEntries(allowedDimensions, "repurposed_structures:well_snow", "lotr:middle_earth");
             addEntries(allowedDimensions, "repurposed_structures:well_mossy_stone", "lotr:middle_earth");
+            configVersion = 3;
         }
 
         if(configVersion == 3){
             fixKeyEntry(allowedDimensions, "repurposed_structures:dungeon_jungle", "repurposed_structures:dungeons_jungle");
             fixKeyEntry(allowedDimensions, "repurposed_structures:warm_land_ruins", "repurposed_structures:ruins_land_warm");
+            configVersion = 4;
         }
 
         if(configVersion == 4){
@@ -105,13 +107,150 @@ public class RSBiomeDimConfig implements Config {
             addEntries(disallowedBiomes, "repurposed_structures:witch_hut_oak", "vampirism:vampire_forest");
             addEntries(disallowedBiomes, "repurposed_structures:outpost_oak", "vampirism:vampire_forest");
             addEntries(disallowedBiomes, "repurposed_structures:well_forest", "vampirism:vampire_forest");
+            configVersion = 5;
         }
 
         if(configVersion == 5){
             addEntries(disallowedDimensions, "all", "thebeginning:.+");
+            configVersion = 6;
         }
 
-        configVersion = 6;
+        if(configVersion == 6) {
+            addEntries(disallowedBiomes, "repurposed_structures:dungeons_badlands", "terralith:snowy_badlands");
+            addEntries(disallowedBiomes, "repurposed_structures:dungeons_snow", "terralith:gravel_desert");
+            addEntries(disallowedBiomes, "repurposed_structures:dungeons_desert", "terralith:red_oasis");
+            addEntries(disallowedBiomes, "repurposed_structures:dungeons_neutral_ocean", "terralith:skylands");
+            addEntries(disallowedBiomes, "repurposed_structures:mansion_desert", "terralith:red_oasis");
+            addEntries(disallowedBiomes, "repurposed_structures:mansion_snowy", "terralith:gravel_desert");
+            addEntries(disallowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:brushland");
+            addEntries(disallowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:volcanic_crater");
+            addEntries(disallowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:volcanic_peaks");
+            addEntries(disallowedBiomes, "repurposed_structures:village_mountains", "terralith:volcanic_crater");
+            addEntries(disallowedBiomes, "repurposed_structures:village_mountains", "terralith:volcanic_peaks");
+            addEntries(disallowedBiomes, "repurposed_structures:village_badlands", "terralith:snowy_badlands");
+            addEntries(disallowedBiomes, "repurposed_structures:outpost_badlands", "terralith:snowy_badlands");
+            addEntries(disallowedBiomes, "repurposed_structures:pyramid_badlands", "terralith:snowy_badlands");
+
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_badlands", "terralith:savanna_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_badlands", "terralith:red_oasis");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_dark_forest", "terralith:mirage_isles");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_desert", "terralith:cave/desert_caves");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_snow", "terralith:alpine_grove");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_snow", "terralith:alpine_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_snow", "terralith:cold_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_icy", "terralith:cave/frostfire_caves");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_icy", "terralith:cave/ice_caves");
+            addEntries(allowedBiomes, "repurposed_structures:dungeons_icy", "terralith:cave/fungal_caves");
+
+            addEntries(allowedBiomes, "repurposed_structures:igloo_grassy", "terralith:brushland");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:amethyst_canyon");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:basalt_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:emerald_peaks");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:granite_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:haze_mountain");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:mountain_steppe");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:painted_mountains");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:rocky_mountains");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:shield_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:steppe");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:stony_spires");
+            addEntries(allowedBiomes, "repurposed_structures:igloo_stone", "terralith:valley_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_birch", "terralith:temperate_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_birch", "terralith:white_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_savanna", "terralith:hot_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_savanna", "terralith:sakura_grove");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_savanna", "terralith:sakura_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_snowy", "terralith:alpine_grove");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_snowy", "terralith:alpine_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_snowy", "terralith:cold_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_snowy", "terralith:snowy_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_snowy", "terralith:snowy_shield");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:forested_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:lavender_forest");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:lavender_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:moonlight_grove");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:moonlight_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:rocky_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:mansion_taiga", "terralith:wintry_forest");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_birch", "terralith:wintry_forest");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_birch", "terralith:white_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_dark_forest", "terralith:mirage_isles");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_desert", "terralith:cave/desert_caves");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_icy", "terralith:cave/frostfire_caves");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_icy", "terralith:cave/ice_caves");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:arid_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:hot_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:sakura_grove");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:sakura_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_savanna", "terralith:savanna_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:amethyst_canyon");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:basalt_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:emerald_peaks");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:granite_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:gravel_desert");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:mountain_steppe");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:steppe");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:shield_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_stone", "terralith:valley_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:cloud_forest");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:forested_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:lavender_forest");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:lavender_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:moonlight_grove");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:moonlight_valley");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:rocky_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:mineshaft_taiga", "terralith:shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_badlands", "terralith:savanna_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_birch", "terralith:temperate_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_birch", "terralith:white_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_snowy", "terralith:alpine_grove");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_snowy", "terralith:alpine_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_snowy", "terralith:cold_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_snowy", "terralith:snowy_shield");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:forested_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:lavender_forest");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:lavender_valley");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:moonlight_grove");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:moonlight_valley");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:rocky_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:outpost_taiga", "terralith:wintry_forest");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_badlands", "terralith:savanna_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_flower_forest", "terralith:blooming_plateau");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_flower_forest", "terralith:blooming_valley");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_snowy", "terralith:alpine_grove");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_snowy", "terralith:alpine_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_snowy", "terralith:cold_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:pyramid_snowy", "terralith:snowy_shield");
+            addEntries(allowedBiomes, "repurposed_structures:village_badlands", "terralith:savanna_badlands");
+            addEntries(allowedBiomes, "repurposed_structures:village_birch", "terralith:temperate_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:village_birch", "terralith:white_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:village_dark_oak", "terralith:mirage_isles");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:amethyst_canyon");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:basalt_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:emerald_peaks");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:granite_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:highlands");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:mountain_steppe");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:rocky_mountains");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:steppe");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:stony_spires");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:shield_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:village_mountains", "terralith:valley_clearing");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_birch", "terralith:temperate_highlands");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_birch", "terralith:white_cliffs");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_dark_forest", "terralith:mirage_isles");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_oak", "terralith:brushland");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:bryce_canyon");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:cloud_forest");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:lavender_forest");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:lavender_valley");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:moonlight_grove");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:moonlight_valley");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:rocky_shrubland");
+            addEntries(allowedBiomes, "repurposed_structures:witch_hut_taiga", "terralith:shrubland");
+        }
+
+        configVersion = 7;
         Config.super.save();
     }
 
@@ -319,5 +458,5 @@ public class RSBiomeDimConfig implements Config {
             "\n"+
             "\n// for internal use only. Do not change this."
     )
-    public int configVersion = 6;
+    public int configVersion = 7;
 }